fix(autoscaler/ecs): change MaxScaleDownPercentage from Int to Float
Eliminate unmarshalling errors by fixing the type of `maxScaleDownPercentage`: `failed to read
group: json: cannot unmarshal number N.N into Go struct field
AutoScaleDown.thirdPartiesIntegration.ecs.autoScale.down.maxScaleDownPercentage of type int`

+++ spotinst/elastigroup_aws_integrations/fields_spotinst_elastigroup_aws_integrations.go
@@ -60,8 +60,8 @@ func expandAWSGroupAutoScaleDown(data interface{}) (*aws.AutoScaleDown, error) {
 				autoScaleDown.SetEvaluationPeriods(spotinst.Int(v))
 			}
 
-			if v, ok := m[string(MaxScaleDownPercentage)].(int); ok && v > 0 {
-				autoScaleDown.SetMaxScaleDownPercentage(spotinst.Int(v))
+			if v, ok := m[string(MaxScaleDownPercentage)].(float64); ok && v > 0 {
+				autoScaleDown.SetMaxScaleDownPercentage(spotinst.Float64(v))
 			}
 		}
 		return autoScaleDown, nil
